loading

+++ src/cli/companion/repos/clean.ts
@@ -1,6 +1,6 @@
 import { readdirs } from "../../../lib/utils";
 import { Config } from "../../../lib/config";
-import { confirm, search } from "../../../lib/ui";
+import { confirm, loading, progressBar, search } from "../../../lib/ui";
 import path from "node:path";
 import { Repo } from "../../../interface/files/repo";
 import { Argv } from "yargs";
@@ -31,34 +31,40 @@ export default {
       const sure = await confirm({ message: `Are you sure you want to clean ${name}` });
       if (!sure) return;
 
+      const spinner = loading(`Cleaning ${name}`);
       await deleteBranches(repo);
+      spinner.succeed(`Succesfully cleaned ${name}`);
 
     } else {
       const sure = await confirm({ message: "Are you sure you want to clean ALL repositories?" });
       if (!sure) return;
 
-      for (const p of paths) {
-        const repo = new Repo(path.join(p.path, p.name));
-        await deleteBranches(repo);
+      const bar = progressBar(paths.length, 0, "cleaning");
+      for (let index = 0; index < paths.length; index += 10) {
+        const toClean = paths.slice(index, index + 10);
+        await Promise.all(toClean.map(async (p) => {
+          bar.setSufix(p.name);
+          const repo = new Repo(path.join(p.path, p.name));
+          await deleteBranches(repo);
+          bar.increment(1);
+        }
+        ));
       }
+      bar.stop();
+      log.success("Succesfully cleaned all repositories");
     }
   }
 };
 
 async function deleteBranches(repo: Repo) {
-  const { name } = await repo.getInfo();
   const to_delete = ["nivelacion", "feature", "bugfix", "hotfix", "fix", "despliegue"];
 
   await repo.stash(async () => {
     await repo.switchBranchIfExists("master");
     const branches = await repo.getBranches();
     for (const branch of branches) {
-      if (to_delete.some((d) => branch.includes(d))) {
-        await repo.deleteBranch(branch);
-        log.info(`Deleting ${branch} of ${name}`);
-      }
+      if (to_delete.some((d) => branch.includes(d))) await repo.deleteBranch(branch);
     }
-    log.success(`Succesfully cleaned ${name}`);
   });
 }
